Reorganize trackers
